pb in dev func print tab free... to be continued

+++ src/realloc.c
@@ -6,15 +6,15 @@
 /*   By: galy <galy@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/01/22 18:34:17 by galy              #+#    #+#             */
-/*   Updated: 2018/01/24 19:24:38 by galy             ###   ########.fr       */
+/*   Updated: 2018/01/26 17:58:56 by galy             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../inc/malloc.h"
+#include "malloc.h"
 
-unsigned int	get_ptr_idx(void *ptr)
+long	get_ptr_idx(void *ptr)
 {
-	unsigned int i;
+	long i;
 
 	i = 0;
 	while (i < vault.meta_items_max)
@@ -28,7 +28,7 @@ unsigned int	get_ptr_idx(void *ptr)
 	return (-1);
 }
 
-void			*go_realloc(size_t size, unsigned int old_idx)
+void			*go_realloc(size_t size, long old_idx)
 {
 	void	*new_ptr;
 	
@@ -54,12 +54,10 @@ void			*realloc(void *ptr, size_t size)
 {
 	// ft_putstr("CALL REALLOC\n");
 
-	unsigned int	old_idx;
-	// void			*old_ptr;
+	long	old_idx;
 	
 	if ((long long)(old_idx = get_ptr_idx(ptr)) == -1)
 		return (NULL);
-	// old_ptr = vault.tab_meta[old_idx].adr;
 	if (vault.tab_meta[old_idx].size == size)
 		return (ptr);
 	else
